requirements.txt
appnope==0.1.0
asgiref==3.3.1
backcall==0.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
django-admin-sortable2==0.7.7
django-cors-headers==3.5.0
django-reverse-admin==2.9.3
djangorestframework==3.12.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.1
wcwidth==0.2.5
